ai-ltd/requirements.txt
streamlit
google-generativeai
python-dotenv
googlesearch-python==1.0.0
google-search-results
torch

pytesseract
langchain
langchain-community
langchain-gemini
pymysql
openai
transformers
flask-session
SpeechRecognition
qrcode[pil]
flask-cors


